% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimStrataMCpbt.R
\name{estimStrataMCpbt}
\alias{estimStrataMCpbt}
\title{This function uses the input created by \code{prepStrata} to estimate the
composistion of each strata.}
\usage{
estimStrataMCpbt(prepData, iter, burnIn, thin = 1, seed = NA)
}
\arguments{
\item{prepData}{The output of \code{prepStrat} with priors and initial values modified as you see fit}

\item{iter}{The total number of iterations to run each strata, including burn-in iterations}

\item{burnIn}{The number of burn-in iterations}

\item{thin}{Thinning parameter for the MCMC chain}

\item{seed}{a positive integer to seed the random number generator. If not specified, chosen based on the current time.}
}
\description{
A wrapper for \code{MCpbt}. This function is used to run the MCMC chain and produce estimates.
}
