% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimStrataPropClip.R
\name{estimStrataPropClip}
\alias{estimStrataPropClip}
\title{This function uses the input created by \code{prepMultStrataPropClip} to estimate the
proportion of fish that are clipped for each strata.}
\usage{
estimStrataPropClip(prepClipData, NumResults, seed = NA, priors = NA,
  verbose = TRUE)
}
\arguments{
\item{prepClipData}{the output of prepMultStrataPropClip}

\item{NumResults}{the number of samples from the posterior to take}

\item{seed}{A positive integer to seed the random number generator. If NA, a seed will
be chosen based on the time.}

\item{priors}{a list of numeric vectors giving the parameters of a beta distribution to use
as the priors for each strata (in order of the strata). The default is a uniform Beta(1,1)
prior for all strata.}

\item{verbose}{TRUE to print some output}
}
\description{
A wrapper for \code{propClipped}
}
