% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generatePBTGSIdata.R
\name{generatePBTGSIdata}
\alias{generatePBTGSIdata}
\title{This function simulates data for one strata.}
\usage{
generatePBTGSIdata(sampRate = 0.1, censusSize = 3000,
  relSizePBTgroups = 1, tagRates = 0.8, obsTagRates = tagRates,
  physTagRates = 0, true_clipped = 0.5, true_noclip_H = 0.2,
  true_wild = 0.3, relSizeGSIgroups = 1, PBT_GSI_calls = 1,
  varMatList = NA)
}
\arguments{
\item{sampRate}{the probability of sampling an individual in the population}

\item{censusSize}{the total size of the populations}

\item{relSizePBTgroups}{a numeric vector giving the relative sizes of the adipose fin intact PBT tagged groups}

\item{tagRates}{a numeric vector giving the true PBT tag rates of the PBT tagged groups, in the
same order as \code{relSizePBTgroups}}

\item{obsTagRates}{a numeric vector giving the PBT tag rates of the PBT tagged groups to use when creating the
tag rate output, in the same order as \code{relSizePBTgroups}}

\item{physTagRates}{a numeric vector giving the true physical tagging rate (CWTs, PIT tags) of the PBT tagged groups, in the
same order as \code{relSizePBTgroups}. Physical tagging is assumed to be independent of PBT tagging}

\item{true_clipped}{the true proportion of the population with a clipped adipose fin}

\item{true_noclip_H}{the true proportion of the population that is hatchery origin and has an intact adipose fin}

\item{true_wild}{the true proportion of the population that is wild origin}

\item{relSizeGSIgroups}{a numeric vector giving the relative sizes of GSI groups within the wild population}

\item{PBT_GSI_calls}{a numeric matrix with rows equal to PBT groups, columns equal to GSI group and
values giving the true proportions of the PBT fish that GSI assign to that GSI group}

\item{varMatList}{a list of numeric matrices, one for each variable. Rows are PBT/GSI groups, in order of PBT groups followed
by GSI groups, columns being the categories for the variable, and values being the proportion of the
group in that category}
}
\description{
To simulate multiple strata, one can run this repeatedly and add a strata variable
}
