% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marginalize_SD.r
\name{marginalize_SD}
\alias{marginalize_SD}
\title{Produce estimates and CIs after marginalizing out a given variable from the output of \code{SCOBI_deux}}
\usage{
marginalize_SD(
  estimatesFile = NA,
  bootHier = NA,
  marginalize = "GenParentHatchery",
  alph = 0.1
)
}
\arguments{
\item{estimatesFile}{is the Scobi-Deux output "Estim_Totals" or "Estim_Grand_Totals" file that you
want to marginalize over.}

\item{bootHier}{is the "Boot_Hier" file for the corresponding variable in your "Estim_Grand_Totals" file.
Note that use of this option requires that you feed the program the "Estim_Grand_Totals" file, not the
"Estim_Totals" file}

\item{marginalize}{is the name of the variable(s) you want to marginalize over. This shoudl be a character
vector and can have one or multiple values.}

\item{alph}{This is the alpha value to use for confidence intervals. So, an alph of .05 will give a
95\% CI, with quantiles at .025 and .975.}
}
\description{
This function requires first running \code{SCOBI_deux} with \code{writeBoot = TRUE}.
An example use is for marginalizing the PBTGroup variable to estimate PBT-derived categories such as PBT-age.
}
\details{
description2
}
