% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MLEwrapper.R
\name{MLEwrapper}
\alias{MLEwrapper}
\title{Find maximum likelihood estimates for composition}
\usage{
MLEwrapper(trapData, tags, GSIcol, PBTcol, strataCol, adFinCol,
  AI = TRUE, optimMethod = "Nelder-Mead", variableCols = NULL, ...)
}
\arguments{
\item{trapData}{a dataframe with a rwo for each individual and columns for GSI assignment, PBT assignment, etc.}

\item{tags}{a dataframe with the first column containing names of PBT groups, and the second column containing
tag rates}

\item{GSIcol}{name of column containing GSI assignments. If you have no GSI information, create a column with
the same value for all samples.}

\item{PBTcol}{name of column containing PBT assignments}

\item{strataCol}{name of column indicating the strata the observation belongs to}

\item{adFinCol}{name of column containing adipose fin status - TRUE (or AI) being intact FALSE (or AD) being clipped, NA missing}

\item{AI}{TRUE to analyze ad-intact fish, FALSE to analyze ad-clipped fish}

\item{optimMethod}{the method to use first with \code{optim}. If this method fails, "Nelder-Mead" is attempted
and a warning is issued.}

\item{variableCols}{name of column containing the variable to estimate composition for (optional)}

\item{...}{other arguments to pass to \code{optim}}
}
\description{
The likelihood is maximized using \code{optim} and some typically close, but quick to calculate, starting values.
If estimating composition of a separate variable, all samples with missing data for this variable must be removed
prior to running this function.
}
