% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multByEscapement.R
\name{multByEscapement}
\alias{multByEscapement}
\title{Multiplies escapement estimates by composition estimates}
\usage{
multByEscapement(prepData, mainRes, clipRes = NA, popSizes,
  verbose = TRUE, writeSummary = TRUE, prefix = "", alpha = 0.1)
}
\arguments{
\item{prepData}{the same \code{prepData} used as input for \code{estimStrataMCpbt}}

\item{mainRes}{the output of \code{estimStrataMCpbt}}

\item{clipRes}{the output of \code{estimStrataPropClip}}

\item{popSizes}{a list of population size estimates for each strata, in order, 
with each strata being an entry in the list}

\item{verbose}{FALSE to repress output written to the screen}

\item{writeSummary}{FALSE to repress summary files being written}

\item{prefix}{the prefix to use when namign the summary files}
}
\description{
Multiplies escapement estimates by composition estimates
}
