% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gradient_negllh_allGSI.R
\name{params_grad}
\alias{params_grad}
\title{this calculates the gradient with respect to the params being optimized when only PBT
and GSI information are used (no other variables)}
\usage{
params_grad(params, nPBT, nGSI, ohnc, t, utGSI, ohnc_gsi)
}
\arguments{
\item{params}{vector of paramaters to optimize}

\item{nPBT}{number of PBT groups to estimate}

\item{nGSI}{number of GSI groups to estimate}

\item{ohnc}{vector of number of observed (PBT assigned) hatchery fish in each PBT and GSI group (gsi groups should be 0)}

\item{t}{vector of tag rates for all PBT and GSI groups (gsi groups should be 0)}

\item{utGSI}{vector of number of un-PBT assigned fish in each GSI group}

\item{ohnc_gsi}{matrix of counts of fish GSI assigned to various groups}
}
\description{
this is not intended for direct user-level use, but it is exported to allow
the user to optionally specify it as a argument to pass to \code{optim}
when using \code{MLEwrapper}. I recommend using it unless you have a very specific
reason not to.
}
\details{
uses the output of test_gradient (chain rule for partial derivatives)
}
