% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gradient_negllh_var.R
\name{params_grad_var}
\alias{params_grad_var}
\title{this calculates the gradient with respect to the params being optimized when PBT,
GSI and another variable are used}
\usage{
params_grad_var(params, nPBT, nGSI, ohnc, t, ohnc_gsi, utVar, ohnc_var, nCat)
}
\arguments{
\item{params}{vector of paramaters to optimize}

\item{nPBT}{number of PBT groups to estimate}

\item{nGSI}{number of GSI groups to estimate}

\item{ohnc}{vector of number of observed (PBT assigned) hatchery fish in each PBT and GSI group (gsi groups should be 0)}

\item{t}{vector of tag rates for all PBT and GSI groups (gsi groups should be 0)}

\item{ohnc_gsi}{matrix of counts of fish GSI assigned to various groups}

\item{utVar}{matrix of numbers of un-PBT assigned fish in each GSI group (row) x category of the variable (column)}

\item{ohnc_var}{matrix of counts of PBT-assigned fish in each category (rows are PBT groups)}

\item{nCat}{the numbers of categories in the variable}
}
\description{
this is not intended for direct user-level use, but it is exported to allow
the user to optionally specify it as a argument to pass to \code{optim}
when using \code{MLEwrapper}. I recommend using it unless you have a specific
reason not to.
}
