% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepStrataPropClip.R
\name{prepStrataPropClip}
\alias{prepStrataPropClip}
\title{This function prepares the data to run multiple strata for estiamting the proportion clipped}
\usage{
prepStrataPropClip(trapData, strataCol, adFinCol, verbose = TRUE)
}
\arguments{
\item{trapData}{dataframe with data for fish sampled from one strata - trap data for dam escapement}

\item{strataCol}{column name of column indicating the strata the observation belongs to}

\item{adFinCol}{column name of column containing adipose fin status - TRUE being intact FALSE being clipped, NA missing}

\item{verbose}{TRUE to print some messages, FALSE to not}
}
\description{
This function prepares the data to run multiple strata for estiamting the proportion clipped
}
