% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rearTypeSim.R
\name{rearTypeSim}
\alias{rearTypeSim}
\title{Performs simulations to estimate precision, bias, and coverage of clipped, noclip_H, and wild estimator}
\usage{
rearTypeSim(numSims, sampRates, censusSizes, relSizePBTgroups, tagRates,
  obsTagRates, physTagRates, true_clipped, true_noclip_H, true_wild,
  alph = 0.1, B = 1000)
}
\arguments{
\item{numSims}{the number of simulations to perform}

\item{sampRates}{sampling rates (dam trap rates) for each strata. Order must match
order of \code{censusSizes}}

\item{censusSizes}{population sizes (dam window counts) for each strata. order must
match order of \code{sampRates}}

\item{relSizePBTgroups}{relative sizes of PBT groups for noclip_H fish ONLY, because
PBT info is not used to determine the proportion clipped}

\item{tagRates}{the true tag rates for each PBT group in the same order as \code{relSizePBTgroups}}

\item{obsTagRates}{the tag rates to use in the estimation for each PBT group in the
same order as \code{relSizePBTgroups}}

\item{physTagRates}{the true physical tagging rate for each PBT groups for noclip_H fish ONLY
in the same order as \code{relSizePBTgroups}}

\item{true_clipped}{the true proportion of fish that are clipped, hatchery origin}

\item{true_noclip_H}{the true proportion of fish that are unclipped, hatchery origin}

\item{true_wild}{the true proportion of fish that are unclipped, wild origin}

\item{alph}{the value of \code{alph} to use for CI estimation}

\item{B}{the number of bootstraps to perform for CI estimation}
}
\value{
a dataframe with estimates and CIs for all three categories from each iteration
}
\description{
Performs simulations to estimate precision, bias, and coverage of clipped, noclip_H, and wild estimator
}
