% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_time_graph.r
\name{run_time_graph}
\alias{run_time_graph}
\title{Making run-timing graphs from output of \code{SCOBI_deux}}
\usage{
run_time_graph(
  windowData = NULL,
  Run = "output",
  dailyCountFile = NULL,
  estimTotalsFile = NULL,
  makePlot = FALSE
)
}
\arguments{
\item{windowData}{This is the exact same \code{windowData} input that you used when you ran \code{SCOBI_deux}.}

\item{Run}{This is the exact same input for \code{Run} you used when you ran \code{SCOBI_deux}. The function looks
for files with this prefix in the working directory to use to make the graph.}

\item{dailyCountFile}{This either a dataframe or it is the path to a csv file that has a list of counts for the days
(or other timestep) that is your basic
unit for graphing. This csv should have two columns, the first column is the group (matching the group - leftmost column - in your
window count file) for that day and the second column is the count on that day. Each consecutive day should be a
new row with the first day being the first row and the last day being the last row. The group will be repeated once for
each day within that group with the count for that particular day.}

\item{estimTotalsFile}{The path to the "Estim_Total" file that you want to break into daily counts and (optionally) make a graph.
This can also be a character vector to multiple files.}

\item{makePlot}{If TRUE, a pdf will be saved in the working directory of a run timing plot.}
}
\description{
This function assumes that proportions or fish of each type are constant within each strata
it is not meant to be a particularly robust method of looking at run timing, rather it is for exploratory use
}
\details{
description2
}
