% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scobi_deux.r
\name{SCOBI_deux}
\alias{SCOBI_deux}
\title{Uses PBT assignments and other data to estimate group composition}
\usage{
SCOBI_deux(adultData = NULL, windowData = NULL, Run = "output",
  RTYPE = "wild", Hierarch_variables = NULL, SizeCut = NULL,
  alph = 0.1, B = 5000, writeBoot = FALSE, pbtRates = NULL,
  adClipVariable = "AdClip", physTagsVariable = "PhysTag",
  pbtGroupVariable = "GenParentHatchery", lengthVariable = "LGDFLmm",
  dataGroupVariable = "WeekNumber", screenOutput = NULL,
  spibetr = TRUE)
}
\arguments{
\item{adultData}{This is the detailed dataset that is used to determine composition. In
dam escapement, this is the "trap" data. It can either be a dataframe or the path to a
csv file with a header.}

\item{windowData}{This is the dataset that gives the census numbers for each group (such
as time period or location). In dam escapement, this is the "window count" dataset.
It can either be a dataframe or the path to a csv file with a header. Format is ....}

\item{Run}{This is a prefix that will be appended to all output files.}

\item{RTYPE}{This is the type of fish that you want to analyze for the \code{Hierarch_variables}.
Options are: "clipped", "noclip_H", "wild", representing Ad-clipped hatchery, Ad-intact
hatchery, and wild fish, respectively.}

\item{Hierarch_variables}{These are the variables (column names in \code{adultData}) that you
want to assess RTYPE for in the order you want them to be assessed.}

\item{SizeCut}{Fish less than this value will be put into a "Small" category, and fish greater or equal
to this value will be put into a "Large" category. This is intended for easy analysis of jacks or
large and small steelhead.}

\item{alph}{This is the alpha value to use for confidence intervals. So, an alph of .05 will give a
95% CI, with quantiles at .025 and .975.}

\item{B}{This is the number of bootstrap iteratiosn to perform. To skip bootstraping and only produce
point estiamtes, use a value of 0.}

\item{writeBoot}{If \code{TRUE}, the values for each bootstrap iteration will be written.}

\item{pbtRates}{This is either a dataframe or the path to a csv file with a header. The first
column contains the names of all the pbtGroupVariable values, and the second column contains their
respective tag rates.}

\item{adClipVariable}{This is the name of the variable that has the ad-clip and ad-intact status of
all samples. Values in this column shoudl either be "AD" or "AI" for ad-clipped and ad-intact,
respectively. NA values are considered missing and are removed from the dataset prior to any analysis.}

\item{physTagsVariable}{This is the name of the variable that has information on physical tags (such
as CWTs). Fish with a physical tag that identifies them as hatchery origin should have the value "tag"
and fish without such a physical tag should have the value "notag". Fish that were not assessed for the
presence of a physical tag should have the value NA.}

\item{pbtGroupVariable}{This is the name of the variable that has the pbt group that a fish assigned to.
Fish that did not assign to any group should be given the value of "Unassigned". Fish that were not
assessed with PBT, for example they failed to genotype or were not genotyped, should be given the value
of NA.}

\item{lengthVariable}{If you are usign the SizeCut option, this is the variable that has the
numeric lengths.}

\item{dataGroupVariable}{This is the name of the variable that has the sample group that each sample
belongs to, with the sample groups matching the groups given in the first column of windowData.}

\item{screenOutput}{This is the name of the file to write all output the would otherwise be printed to
the console. If not specified, output is printed to the console.}

\item{spibetr}{If this is \code{TRUE}, then the composition estimates of the wild fish will be corrected
to account for the expected number and composition of the unclipped, untagged hatchery fish present in
the wild sample.}

\item{lengthVariable}{This is the name of the variable that has numberic lengths if the SizeCut option
is being used.}
}
\value{
A string is returned simply stating that the analysis is complete. All of
the output of the function is written to the working directory. These files are:
....
....
....
}
\description{
\code{SCOBI_deux} Uses PBT assignments and tag rates to calculate composition
estimates and adjust estimates based on the tag rates. Confidence intervals are
generated by non-parametric bootstrapping.
}
\details{
\code{SCOBI_deux} Uses PBT assignments and tag rates to calculate composition
estimates and adjust estimates based on the tag rates. Confidence intervals are
generated by non-parametric bootstrapping. This function... further explanation
...
...
}
