% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/old_approach_hier_boot_scobi_no_pbt.R,
%   R/scobi_no_pbt.R
\name{scobiNoPBT}
\alias{scobiNoPBT}
\title{generates estiamtes and CIs usings census count and trap data without any PBT}
\usage{
scobiNoPBT(adultData = NULL, windowData = NULL, Run = "output",
  RTYPE = "wild", Hierarch_variables = NULL, SizeCut = NULL,
  alph = 0.1, B = 5000, writeBoot = FALSE,
  adClipVariable = "AdClip", physTagsVariable = "PhysTag",
  lengthVariable = "LGDFLmm", dataGroupVariable = "WeekNumber")

scobiNoPBT(adultData = NULL, windowData = NULL, Run = "output",
  RTYPE = "wild", Hierarch_variables = NULL, SizeCut = NULL,
  alph = 0.1, B = 5000, writeBoot = FALSE,
  adClipVariable = "AdClip", physTagsVariable = "PhysTag",
  lengthVariable = "LGDFLmm", dataGroupVariable = "WeekNumber")
}
\value{
something

something
}
\description{
\code{scobiNoPBT} generates estiamtes and CIs usings census count and trap data
without any PBT data. This function uses parametric bootstrapping to generate CIs

\code{scobiNoPBT} generates estiamtes and CIs usings census count and trap data
without any PBT data. This function uses parametric bootstrapping to generate CIs
}
\details{
Some longer explanation here... someday... probably right before publishing...

Some longer explanation here... someday... probably right before publishing...
}
